@@ -36,7 +36,7 @@ final class MercanetBnpParibasGatewayFactory extends GatewayFactory
             $config->defaults($config['payum.default_options']);
             $config['payum.required_options'] = ['secret_key', 'environment', 'merchant_id', 'key_version'];
 
-            $config['payum.api'] = function (ArrayObject $config) {
+            $config['payum.api'] = function (ArrayObject $config): Bridge\MercanetBnpParibasBridgeInterface {
                 $config->validateNotEmpty($config['payum.required_options']);
 
                 /** @var MercanetBnpParibasBridgeInterface $mercanetBnpParibasBridge */
